+++ packages/authenticated-user-storage/tests/fixtures/authenticated-userstorage.ts
@@ -11,6 +11,8 @@ import {
   MOCK_IDENTITY_SHARING_CONSENT_URL,
   MOCK_NOTIFICATION_PREFERENCES,
   MOCK_NOTIFICATION_PREFERENCES_URL,
+  MOCK_USER_ASSETS_BLOB,
+  MOCK_USER_ASSETS_URL,
 } from '../mocks/authenticated-userstorage.js';
 
 type MockReply = {
@@ -155,6 +157,32 @@ export function handleMockSetAssetsWatchlist(
 ): nock.Scope {
   const reply = mockReply ?? { status: 200 };
   const interceptor = nock(MOCK_ASSETS_WATCHLIST_URL).persist().put('');
+
+  if (callback) {
+    return interceptor.reply(reply.status, async (uri, requestBody) => {
+      return callback(uri, requestBody);
+    });
+  }
+  return interceptor.reply(reply.status, reply.body);
+}
+
+export function handleMockGetUserAssets(mockReply?: MockReply): nock.Scope {
+  const reply = mockReply ?? {
+    status: 200,
+    body: MOCK_USER_ASSETS_BLOB,
+  };
+  return nock(MOCK_USER_ASSETS_URL)
+    .persist()
+    .get('')
+    .reply(reply.status, reply.body);
+}
+
+export function handleMockSetUserAssets(
+  mockReply?: MockReply,
+  callback?: (uri: string, requestBody: nock.Body) => Promise<void>,
+): nock.Scope {
+  const reply = mockReply ?? { status: 200 };
+  const interceptor = nock(MOCK_USER_ASSETS_URL).persist().put('');
 
   if (callback) {
     return interceptor.reply(reply.status, async (uri, requestBody) => {